+++ Editor/PreviewSystem/IRenderFilter.cs
@@ -117,8 +117,18 @@ namespace nadena.dev.ndmf.preview
         /// <param name="context">A compute context that is used to track which values your code depended on in
         ///     configuring this node. Changing these values will triger a recomputation of this node.</param>
         /// <returns></returns>
-        public Task<IRenderFilterNode> Instantiate(RenderGroup group, IEnumerable<(Renderer, Renderer)> proxyPairs,
-            ComputeContext context);
+        public Task<IRenderFilterNode> Instantiate(
+            RenderGroup group,
+            IEnumerable<(Renderer, Renderer)> proxyPairs,
+            ComputeContext context
+        );
+
+        // Allow for future expansion
+        [ExcludeFromDocs]
+        [UsedImplicitly]
+        void __please_enable_dotnet_80_or_higher_for_default_methods()
+        {
+        }
     }
 
     [Flags]
@@ -148,12 +158,6 @@ namespace nadena.dev.ndmf.preview
     public interface IRenderFilterNode : IDisposable
     {
         /// <summary>
-        /// Indicates which static aspects of a renderer this node examines. Changes to these aspects will trigger a
-        /// rebuild or partial update of this node.
-        /// </summary>
-        public RenderAspects Reads { get; }
-
-        /// <summary>
         /// Indicates which aspects of a renderer this node changed, relative to the node prior to the last Update
         /// call. This may trigger updates of downstream nodes.
         ///
@@ -182,6 +186,7 @@ namespace nadena.dev.ndmf.preview
         /// </summary>
         /// <param name="proxyPairs"></param>
         /// <param name="context"></param>
+        /// <param name="renderFilterContext"></param>
         /// <param name="updatedAspects"></param>
         /// <returns></returns>
         public Task<IRenderFilterNode> Refresh(
@@ -208,6 +213,13 @@ namespace nadena.dev.ndmf.preview
         }
 
         void IDisposable.Dispose()
+        {
+        }
+
+        // Allow for future expansion
+        [ExcludeFromDocs]
+        [UsedImplicitly]
+        void __please_enable_dotnet_80_or_higher_for_default_methods()
         {
         }
     }