+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [#287] Added PublishedValue class
+- [#288] Added support for passing ObjectRegistry to IRenderFilter
 
 ### Fixed
 - [#283] Cached proxy objects are visible after exiting play mode